优化：开关拨动的时候，PWM可以马上响应
之前是PWM输出过程中，拨动开关，必须要之前模式下的PWM输出完毕才能切换到当前模式的PWM；现在优化了，开关一拨动，PWM马上根据当前开关的档位进行切换；
具体做法：增加了一个任务，20ms，检测开关状态，一旦发现开关换挡了，马上换对应的PWM

+++ EAR_F030/MODULE/protocol_module.c
@@ -439,7 +439,7 @@ uint8_t FlashReadByte(uint32_t addr)
 }
 
 //得到按键模式
-uint16_t GetModeSelected(void)
+uint8_t GetModeSelected(void)
 {
 	uint16_t res;
 	res=RegularConvData_Tab[1];
